feat: moved signIn method to User domain
