+++ src/main/java/com/lazyworking/sagupalgu/user/service/UserService.java
@@ -1,5 +1,10 @@
 package com.lazyworking.sagupalgu.user.service;
 
+import com.lazyworking.sagupalgu.login.form.SignInForm;
+import com.lazyworking.sagupalgu.resources.domain.Role;
+import com.lazyworking.sagupalgu.resources.domain.RoleUser;
+import com.lazyworking.sagupalgu.resources.repository.RoleRepository;
+import com.lazyworking.sagupalgu.resources.repository.RoleUserRepository;
 import com.lazyworking.sagupalgu.user.domain.Gender;
 import com.lazyworking.sagupalgu.user.domain.ReportedUsers;
 import com.lazyworking.sagupalgu.user.domain.User;
@@ -10,6 +15,7 @@ import com.lazyworking.sagupalgu.user.repository.ReportedUsersRepository;
 import com.lazyworking.sagupalgu.user.repository.UserRepository;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,15 +30,25 @@ public class UserService {
     private final UserRepository userRepository;
     private final ReportedUsersRepository reportedUsersRepository;
 
-    @PostConstruct
-    void initData() {
-        User user1= new User("user1","user@email.com","12345678910", LocalDateTime.now(),Gender.M);
-        User user2= new User("user2","user2@email.com","12345678911", LocalDateTime.now(),Gender.F);
+    private final RoleRepository roleRepository;
 
-        userRepository.save(user1);
-        userRepository.save(user2);
+    private final RoleUserRepository roleUserRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    @Transactional
+    public Long addUser(SignInForm form){
+        User newUser = form.toEntity();
+        newUser.setPassword(passwordEncoder.encode(form.getPassword()));
+        checkDuplicateEmail(form.getEmail());
+        newUser = userRepository.save(newUser);
+        return newUser.getId();
     }
-
+    //중복 이메일 확인
+    private void checkDuplicateEmail(String email) {
+        User user = userRepository.findByEmail(email);
+        if(user != null)
+            throw new IllegalStateException("이미 등록된 회원입니다");
+    }
 
     //회원 정보 수정
     @Transactional
